add:post_moderation_concern.rbを作成しFatController抑制
